Extract PokemonIllustration component
